+++ docs/news.rst
@@ -26,6 +26,8 @@ Psycopg 3.0.10 (unreleased)
 
 - Leave the connection working after interrupting a query with Ctrl-C
   (currently only for sync connections, :ticket:`#231`).
+- Fix `Cursor.description` after a COPY ... TO STDOUT operation
+  (:ticket:`#235`).
 
 
 Current release